clean up aws
remove cli, added boto3
