+++ griptape/tools/__init__.py
@@ -3,23 +3,28 @@ from .web_search.tool import WebSearch
 from .web_scraper.tool import WebScraper
 from .sql_client.tool import SqlClient
 from .email_client.tool import EmailClient
-from .aws_cli.tool import AwsCli
 from .rest_api_client.tool import RestApiClient
 from .file_manager.tool import FileManager
 from .knowledge_base_client.tool import KnowledgeBaseClient
 from .date_time.tool import DateTime
-from .aws_s3.tool import AwsS3
+
+# AWS Tools
+from .base_aws_client import BaseAwsClient
+from .aws_s3_client.tool import AwsS3Client
+from .aws_pricing_client.tool import AwsPricingClient
 
 __all__ = [
+    # AWS
+    "BaseAwsClient",
+    "AwsPricingClient",
+    "AwsS3Client",
     "Calculator",
     "WebSearch",
     "WebScraper",
     "SqlClient",
     "EmailClient",
-    "AwsCli",
     "RestApiClient",
     "FileManager",
     "KnowledgeBaseClient",
-    "DateTime",
-    "AwsS3"
+    "DateTime"
 ]
