--- griptape/tools/aws_cli/requirements.txt
@@ -1,1 +1,0 @@
-git+https://github.com/aws/aws-cli.git@2.11.12#egg=awscli
