--- griptape/tools/aws_s3/requirements.txt
@@ -1,2 +1,0 @@
-json
-boto3
