+++ griptape/tools/aws_s3_client/manifest.yml
@@ -1,5 +1,5 @@
 version: "v1"
-name: AWS CLI
-description: Tool for accessing AWS CLI v2 commands.
+name: AWS S3 Client
+description: Tool for the S3 boto3 API.
 contact_email: hello@griptape.ai
 legal_info_url: https://www.griptape.ai/legal
