+++ griptape/tools/base_aws_client.py
@@ -1,0 +1,22 @@
+from abc import ABC
+import boto3
+from attr import define, field
+from griptape.artifacts import TextArtifact, ErrorArtifact, BaseArtifact
+from griptape.core import BaseTool
+from griptape.core.decorators import activity
+
+
+@define
+class BaseAwsClient(BaseTool, ABC):
+    session: boto3.session = field(kw_only=True)
+
+    @activity(config={
+        "description": "Returns the account and IAM principal of the AWS credentials being used."
+    })
+    def get_current_aws_identity(self, params: dict) -> BaseArtifact:
+        try:
+            session = self.session
+            sts = session.client('sts')
+            return TextArtifact(str(sts.get_caller_identity()))
+        except Exception as e:
+            return ErrorArtifact(f"error getting current aws caller identity: {e}")
